test/system/requirements.txt
argparse
pathlib
pyserial
pyyaml
